pyproject.toml
[project]
name = "pyxed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.0.0",
    "rich>=14.2.0",
]
